reporting-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
paho-mqtt>=1.6.0
python-dotenv>=1.0.0
